requirements.txt
garminconnect==0.1.55
python-dotenv==1.0.0
psycopg2-binary==2.9.9
schedule==1.2.1
pandas==2.1.4
numpy==1.26.3
streamlit==1.32.0
plotly==5.18.0
sqlalchemy==2.0.27
scipy==1.13.0
